feat(cmd/gpud): separate dir for "gpud login"
